+++ src/discljord/connections/specs.clj
@@ -41,3 +41,10 @@
 (s/def ::gateway (s/keys :req-un [::url ::shard-count ::session-start-limit]))
 
 (s/def ::identify-when ifn?)
+(s/def ::disable-compression boolean?)
+
+(s/def :shard/id nat-int?)
+(s/def :shard/session string?)
+(s/def :shard/count pos-int?)
+(s/def :shard/seq nat-int?)
+(s/def ::shard (s/keys :req-un [:shard/id :shard/session :shard/count :shard/seq]))